requirements.txt

streamlit
httpx
openai
python-dotenv
groq
httpx
fastapi
uvicorn




